refactor: move upd tracker client
to follow same folder strucutre as the HTTP tracker client.
